USNG-3 Fix distance bug edge case
USNG boxes are squares with sides of multiples of 10, starting with 0.
We use the diagonal distance between lat/long points to calculate the precision
that the box should be. This was causing an edge case with a USNG box of size 10
for example would give a lat/long distance of 14. When recalculating the USNG box,
this would turn into a box of size 100. We now use trigonometry to
calculate the side length instead of using the diagonal distance.