USNG-3 Fix intermittent northing issue
Northing was calculated incorrectly for certain bounding boxes.
In alternating horizontal bands, when drawing the largest zone, e.g,
12R, 13R, 17M, etc, the drawing would jump up a region. I.e, 12R would
appear where 12S is. 12S would still be in the correct place.
This seems to have been an edge case condition with the largest regions
having northings ignored. Now, when a region is the largest, it has approx.
500k added to its northing to estimate the middle and more accurately
represent its position.